+++ semver-gradle-plugin/test/kotlin/com/javiersc/semver/gradle/plugin/GitHubVariablesTest.kt
@@ -12,6 +12,15 @@ class GitHubVariablesTest : GradleTest() {
     @Test
     fun `GitHub variables`() {
         gradleTestKitTest("github-variables") {
+            val githubEnvPath = projectDir.resolve("environment/github.env").path
+            val githubEnvFile =
+                projectDir.resolve(githubEnvPath).apply {
+                    parentFile.mkdirs()
+                    createNewFile()
+                }
+
+            withEnvironment(mapOf("GITHUB_ENV" to githubEnvPath))
+
             projectDir.generateInitialCommitAddVersionTagAndAddNewCommit()
 
             gradlew(
@@ -22,7 +31,9 @@ class GitHubVariablesTest : GradleTest() {
                     "--githubOutputTag",
                     "--githubOutputVersion",
                     "--githubOutput",
-                    stacktrace(),
+                    "--githubEnvTag",
+                    "--githubEnvVersion",
+                    "--githubEnv",
                 )
                 .outputTrimmed
                 .shouldContain("Setting v as `semver-tag` output:")
@@ -31,6 +42,13 @@ class GitHubVariablesTest : GradleTest() {
                 .shouldContain("::set-output name=semver-version::0.9.1")
                 .shouldContain("Setting v0.9.1 as `semver` output:")
                 .shouldContain("::set-output name=semver::v0.9.1")
+
+            githubEnvFile
+                .readText()
+                .shouldContain("SEMVER-TAG=v")
+                .shouldContain("SEMVER-VERSION=0.9.1")
+                .shouldContain("SEMVER=v0.9.1")
+
             gradlew(
                     "printSemver",
                     "-Psemver.tagPrefix=w",
@@ -38,7 +56,9 @@ class GitHubVariablesTest : GradleTest() {
                     "--githubOutputTag",
                     "--githubOutputVersion",
                     "--githubOutput",
-                    stacktrace(),
+                    "--githubEnvTag",
+                    "--githubEnvVersion",
+                    "--githubEnv",
                 )
                 .outputTrimmed
                 .shouldContain("Setting v as `semver-tag` output:")
@@ -53,6 +73,12 @@ class GitHubVariablesTest : GradleTest() {
                 .shouldContain("::set-output name=semver-version-library::0.1.1")
                 .shouldContain("Setting w0.1.1 as `semver-library` output:")
                 .shouldContain("::set-output name=semver-library::w0.1.1")
+
+            githubEnvFile
+                .readText()
+                .shouldContain("SEMVER-TAG-LIBRARY=w")
+                .shouldContain("SEMVER-VERSION-LIBRARY=0.1.1")
+                .shouldContain("SEMVER-LIBRARY=w0.1.1")
         }
     }
 }