+++ semver-gradle-plugin/test/kotlin/com/javiersc/semver/gradle/plugin/setup/Sandbox.kt
@@ -68,6 +68,7 @@ internal fun File.createGitIgnore() {
                 |build/
                 |.gradle/
                 |local.properties
+                |environment/github.env
                 |
             """.trimMargin()
         )